Add support for picture that are not 28x28
This resizes the picture to 28x28. A side effect is that if the
original picture is not a square, the image will be deformed.

A better approach would be to add a layer to the neural network to
detect the digit on the picture, crop it as a square and make it
28x28. But this is beyond the scope of this test.

This just aim not to raise exceptions.